requirements.txt
langchain-groq
langgraph
langchain-core
python-dotenv
requests
pydantic 